fix(snapshot,event): carry MainSpeed as its raw operand so the cue enums stay Eq
CutsceneCue dropped its Eq derive because the new ActorMove carried an f32
speed, and an extension surface consumers match on must not lose a derive.
The float was never the authored value: 0x32 MainSpeed writes a work slot
that the VM scaled by EVENT_SPEED_SCALE on the way out, so the cue now
carries the raw i32 operand exactly as x/y/z/heading already do, and the
renderer and the MOVE-hold estimator apply the scale (now pub, imported
from the emitter rather than re-derived). Eq returns to EventCue,
CutsceneCue and ResolvedCue; MapOp gains it for the same reason.

kuluu-snapshot also drops its ffxi-proto dependency, which no code used -
every mention was in a doc comment - and which cost the wire crate its
transport-agnostic standing.

The PROTOCOL_VERSION log had been renumbered rather than appended: main
v27/v28 were rewritten to v34/v35 while new v27/v28 entries were minted
for a different field, v32/v33 were appended below v4, and nothing at all
was logged for self_pet_targid, the three DialogState fields, HudHide,
ClockHold or the Map events. Published numbers are restored, the branch
entries renumbered above them, and the unlogged surface written down.

+++ kuluu-session/src/state.rs
@@ -1028,9 +1028,8 @@ pub enum CutsceneActor {
 
 /// One staging effect the running event script asked for — the renderer-facing
 /// half of [`ffxi_event::EventCue`]. `MusicVolume` is absent because 0x5D rides
-/// [`AgentEvent::MusicVolumeChanged`] instead of the cue stream. Not `Eq`:
-/// [`CutsceneCue::ActorMove`] carries a float speed.
-#[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
+/// [`AgentEvent::MusicVolumeChanged`] instead of the cue stream.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 pub enum CutsceneCue {
     ActorMotion {
         actor: CutsceneActor,
@@ -1091,7 +1090,9 @@ pub enum CutsceneCue {
         y: i32,
         z: i32,
         heading: i32,
-        speed: f32,
+        /// Raw 0x32 MainSpeed operand; the renderer scales it with
+        /// `ffxi_event::vm::scene::EVENT_SPEED_SCALE`.
+        speed: i32,
     },
     /// Snap `actor` to `(x, y, z)` facing `heading`, in event-coordinate
     /// integers.
